:recycle: refactor: Change code for to use ASWebAuthenticationSession
Change class of the auth to adapt with documentation

What was changed in this Pull Request?

# Demand

What demand does this Pull Request refer to?

- [ ] Bug
- [ ] Functionality
- [X] update

# Automated tests

- [ ] Unit tests
- [ ] Component Tests
- [ ] Integration tests

Is there any place where the test was not carried out, for what reason?

# Feature toggle

Does demand have a feature toggle? What is the name of the feature?
- [X] Yes
- [ ] No

What is the name of the feature?

ASWeb Authentication

# Guidelines for the evaluator

Does the Pull Request evaluator need to know any more relevant information?